requirements.txt
pyzt==0.0.2
pytz==2023.3
typer==0.9.0
typing-extensions==4.8.0
sqlparse==0.4.4
asgiref==3.7.2
click==8.1.7

django==5.0
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.5

# django-rest-swagger==2.2.0
# coreapi==2.3.3
# openapi-codec==1.3.2
# simplejson==3.19.2
# requests==2.31.0
uritemplate==4.1.1
# coreschema==0.0.4
# itypes==1.2.0
# jinja2==3.1.2
# urllib3==2.1.0
# certifi==2023.11.17
# charset-normalizer==3.3.2
# idna==3.6
# MarkupSafe==2.1.3

drf-yasg==1.21.7
pyyaml==6.0.1
inflection==0.5.1
packaging==23.2

psycopg2==2.9.9